Inclusão login com email e senha via Firebase. Adequação layout tela principal e de sign up
